Migrate project to IntelliJ Idea IDE
